code cleanup +resolved ts errors + description for mediaconstraints
