+++ development/video-recorder-app/components/video-recorder/autoStop/autoStop.jsx
@@ -67,7 +67,7 @@ const AutoStopVideoRecorder = () => {
         </button>
       )}
       {recordingState === "recorded" && (
-        <button className="bg-blue-300 border -2" onClick={downloadHandler}>
+        <button className="bg-blue-300 border-2" onClick={downloadHandler}>
           Download Video in Parent
         </button>
       )}
